Port to new script variable API
