+++ include/flecs/addons/script.h
@@ -10,7 +10,7 @@
 /**
  * @defgroup c_addons_script Flecs script
  * @ingroup c_addons
- * Data definition format for loading entity data.
+ * DSL for loading scenes, assets and configuration.
  *
  * @{
  */
@@ -35,14 +35,71 @@ extern ECS_COMPONENT_DECLARE(EcsScript);
 
 typedef struct ecs_script_t ecs_script_t;
 typedef struct ecs_script_assembly_t ecs_script_assembly_t;
-typedef struct ecs_script_vars_t ecs_script_vars_t;
-typedef struct ecs_script_var_t ecs_script_var_t;
+
+/** Script variable. */
+typedef struct ecs_script_var_t {
+    const char *name;
+    ecs_value_t value;
+    const ecs_type_info_t *type_info;
+} ecs_script_var_t;
+
+/** Script variable scope. */
+typedef struct ecs_script_vars_t {
+    struct ecs_script_vars_t *parent;
+    ecs_hashmap_t var_index;
+    ecs_vec_t vars;
+
+    struct ecs_stack_t *stack;
+    ecs_stack_cursor_t *cursor;
+    ecs_allocator_t *allocator;
+} ecs_script_vars_t;
 
 /* Script component */
 typedef struct EcsScript {
     ecs_script_t *script;
     ecs_script_assembly_t *assembly; /* Only set for assembly scripts */
 } EcsScript;
+
+
+/* Parsing & running scripts */
+
+/** Parse script.
+ * This operation parses a script and returns a script object upon success. To
+ * run the script, call ecs_script_eval().
+ * 
+ * @param world The world.
+ * @param name Name of the script (typically a file/module name).
+ * @param code The script code.
+ * @return Script object if success, NULL if failed.
+*/
+FLECS_API
+ecs_script_t* ecs_script_parse(
+    ecs_world_t *world,
+    const char *name,
+    const char *code);
+
+/** Evaluate script.
+ * This operation evaluates (runs) a script.
+ * 
+ * @param script The script.
+ * @return Zero if success, non-zero if failed.
+*/
+FLECS_API
+int ecs_script_eval(
+    ecs_script_t *script);
+
+/** Free script.
+ * This operation frees a script object.
+ * 
+ * Assemblies created by the script rely upon resources in the script object,
+ * and for that reason keep the script alive until all assemblies created by the
+ * script are deleted.
+ *
+ * @param script The script.
+ */
+FLECS_API
+void ecs_script_free(
+    ecs_script_t *script);
 
 /** Parse script.
  * This parses a script and instantiates the entities in the world.
@@ -72,34 +129,39 @@ int ecs_script_run(
  *
  * @param world The world.
  * @param filename The script file name.
- * @return Zero if success, non-zero otherwise.
+ * @return Zero if success, non-zero if failed.
  */
 FLECS_API
 int ecs_script_run_file(
     ecs_world_t *world,
     const char *filename);
 
-
-
+/** Convert script AST to string.
+ * This operation converts the script abstract syntax tree to a string, which
+ * can be used to debug a script.
+ * 
+ * @param script The script.
+ * @param buf The buffer to write to.
+ * @return Zero if success, non-zero if failed.
+ */
 FLECS_API
-ecs_script_t* ecs_script_parse(
-    ecs_world_t *world,
-    const char *name,
-    const char *code);
+int ecs_script_to_buf(
+    ecs_script_t *script,
+    ecs_strbuf_t *buf);
 
+/** Convert script AST to string.
+ * This operation converts the script abstract syntax tree to a string, which
+ * can be used to debug a script.
+ * 
+ * @param script The script.
+ * @return The string if success, NULL if failed.
+ */
 FLECS_API
-int ecs_script_eval(
+char* ecs_script_to_str(
     ecs_script_t *script);
 
-FLECS_API
-void ecs_script_free(
-    ecs_script_t *script);
 
-
-
-
-
-
+/* Managed scripts */
 
 /** Used with ecs_script_init() */
 typedef struct ecs_script_desc_t {
@@ -132,16 +194,13 @@ ecs_entity_t ecs_script_init(
  * @param script The script entity.
  * @param instance An assembly instance (optional).
  * @param str The script code.
- * @param vars Optional preset variables for script parameterization.
  */
 FLECS_API
 int ecs_script_update(
     ecs_world_t *world,
     ecs_entity_t script,
     ecs_entity_t instance,
-    const char *str,
-    ecs_vars_t *vars);
-
+    const char *str);
 
 /** Clear all entities associated with script.
  *
@@ -156,42 +215,65 @@ void ecs_script_clear(
     ecs_entity_t instance);
 
 
+/* Script variables */
 
-
-
-
-
-
-FLECS_API
-int ecs_script_to_buf(
-    ecs_script_t *script,
-    ecs_strbuf_t *buf);
-
-FLECS_API
-char* ecs_script_to_str(
-    ecs_script_t *script);
-
-
-
-
-
-
+/** Push new variable scope.
+ * A variable scope contains one or more variables. Scopes can be nested, which
+ * allows variables in different scopes to have the same name. Variables from
+ * parent scopes will be shadowed by variables in child scopes with the same
+ * name.
+ * 
+ * Scopes created with ecs_script_vars_push() must be cleaned up with
+ * ecs_script_vars_pop().
+ *
+ * @param parent The parent scope (provide NULL for root scope).
+ * @param stack The stack allocator to use.
+ * @param allocator The allocator to use.
+ * @return The new variable scope.
+ */
 ecs_script_vars_t* ecs_script_vars_push(
     ecs_script_vars_t *parent,
     struct ecs_stack_t *stack,
     ecs_allocator_t *allocator);
 
+/** Pop variable scope.
+ * This frees up the resources for a variable scope. The scope must be at the
+ * top of a vars stack. Calling ecs_script_vars_pop() on a scope that is not the
+ * last scope causes undefined behavior.
+ *
+ * @param vars The scope to free.
+ * @return The parent scope.
+ */
 ecs_script_vars_t* ecs_script_vars_pop(
     ecs_script_vars_t *vars);
 
+/** Declare a variable.
+ * This operation declares a new variable in the current scope. If a variable
+ * with the specified name already exists, the operation will fail.
+ * 
+ * @param vars The variable scope.
+ * @param name The variable name.
+ * @return The new variable, or NULL if the operation failed.
+ */
 ecs_script_var_t* ecs_script_vars_declare(
     ecs_script_vars_t *vars,
     const char *name);
 
+/** Lookup a variable.
+ * This operation looks up a variable in the current scope. If the variable 
+ * can't be found in the current scope, the operation will recursively search
+ * the parent scopes.
+ * 
+ * @param vars The variable scope.
+ * @param name The variable name.
+ * @return The variable, or NULL if a one with the provided name does not exist.
+ */
 ecs_script_var_t* ecs_script_vars_lookup(
     const ecs_script_vars_t *vars,
     const char *name);
 
+
+/* Script utilities */
 
 /** Used with ecs_parse_expr(). */
 typedef struct ecs_parse_expr_desc_t {
@@ -226,10 +308,61 @@ const char* ecs_parse_expr(
     ecs_value_t *value,
     const ecs_parse_expr_desc_t *desc);
 
+/** Evaluate interpolated expressions in string.
+ * This operation evaluates expressions in a string, and replaces them with
+ * their evaluated result. Supported expression formats are:
+ *  - $variable_name
+ *  - {expression}
+ *
+ * The $, { and } characters can be escaped with a backslash (\).
+ *
+ * @param world The world.
+ * @param str The string to evaluate.
+ * @param vars The variables to use for evaluation.
+ */
+FLECS_API
+char* ecs_interpolate_string(
+    ecs_world_t *world,
+    const char *str,
+    const ecs_script_vars_t *vars);
 
-
-
-
+/** Convert iterator to vars
+ * This operation converts an iterator to a variable array. This allows for
+ * using iterator results in expressions. The operation only converts a
+ * single result at a time, and does not progress the iterator.
+ *
+ * Iterator fields with data will be made available as variables with as name
+ * the field index (e.g. "$1"). The operation does not check if reflection data
+ * is registered for a field type. If no reflection data is registered for the
+ * type, using the field variable in expressions will fail.
+ *
+ * Field variables will only contain single elements, even if the iterator
+ * returns component arrays. The offset parameter can be used to specify which
+ * element in the component arrays to return. The offset parameter must be
+ * smaller than it->count.
+ *
+ * The operation will create a variable for query variables that contain a
+ * single entity.
+ *
+ * The operation will attempt to use existing variables. If a variable does not
+ * yet exist, the operation will create it. If an existing variable exists with
+ * a mismatching type, the operation will fail.
+ *
+ * Accessing variables after progressing the iterator or after the iterator is
+ * destroyed will result in undefined behavior.
+ *
+ * If vars contains a variable that is not present in the iterator, the variable
+ * will not be modified.
+ *
+ * @param it The iterator to convert to variables.
+ * @param vars The variables to write to.
+ * @param offset The offset to the current element.
+ */
+FLECS_API
+void ecs_iter_to_vars(
+    const ecs_iter_t *it,
+    ecs_script_vars_t *vars,
+    int offset);
 
 /* Module import */
 FLECS_API
