+++ src/addons/alerts.c
@@ -385,8 +385,9 @@ void MonitorAlertInstances(ecs_iter_t *it) {
         ranges = ecs_ref_get(world, &alert->ranges, EcsMemberRanges);
     }
 
-    ecs_vars_t vars = {0};
-    ecs_vars_init(world, &vars);
+    ecs_script_vars_t *vars = ecs_script_vars_push(NULL,
+        flecs_stage_get_stack_allocator(it->world),
+        flecs_stage_get_allocator(it->world));
 
     int32_t i, count = it->count;
     for (i = 0; i < count; i ++) {
@@ -449,9 +450,9 @@ void MonitorAlertInstances(ecs_iter_t *it) {
                         ecs_os_free(alert_instance[i].message);
                     }
 
-                    ecs_iter_to_vars(&rit, &vars, 0);
+                    ecs_iter_to_vars(&rit, vars, 0);
                     alert_instance[i].message = ecs_interpolate_string(
-                        world, alert->message, &vars);
+                        world, alert->message, vars);
                 }
 
                 if (timeout) {
@@ -472,7 +473,6 @@ void MonitorAlertInstances(ecs_iter_t *it) {
         }
 
         /* Alert instance is no longer active */
-
         if (timeout) {
             if (ECS_EQZERO(timeout[i].inactive_time)) {
                 /* The alert just became inactive. Add Disabled tag */
@@ -494,7 +494,7 @@ void MonitorAlertInstances(ecs_iter_t *it) {
         ecs_delete(world, ai);
     }
 
-    ecs_vars_fini(&vars);
+    ecs_script_vars_pop(vars);
 }
 
 ecs_entity_t ecs_alert_init(
